requirements.txt
prometheus-client==0.20.0
pyserial==3.5
PyYAML==6.0.1
